#47 The profile and the activity viewscoped Beans are ready to get the actually logged in agent's open conversation

+++ helpdesk-service-api/src/main/java/hu/schonherz/training/helpdesk/service/api/vo/ConversationVO.java
@@ -11,11 +11,12 @@ import lombok.NoArgsConstructor;
 @Builder
 public class ConversationVO {
     private long id;
+
     private String clientId;
 
     private String clientEmail;
 
-    private int agentId;
+    private Long agentId;
 
     private boolean closed;
 }
